Both previous and refactored commands now work
